my_maturin_library/requirements.txt
Pillow==10.0.1
Pillow==10.3.0
